+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Improvements
 - Improved ElementHelper.AddTooltip overloads
 - Don't query a paragraph's text callback in the constructor
 - Allow manually hiding a paragraph without its text overriding the hidden state
+- Added optional isKeybindAllowed parameter to KeybindButton
 
 Fixes
 - Fixed auto-nav tooltip displaying on the selected element even when not in auto-nav mode